+++ src/Components/SingleMovie/SingleMovie.js
@@ -2,19 +2,26 @@ import './SingleMovie.css';
 import backArrow from './backArrow.png';
 
 const SingleMovie = (props) => {
+    console.log(props.singleMovie.backdrop_path)
+    const background = {
+        backgroundImage: `url(${props.singleMovie.backdrop_path})`,
+        backgroundSize: 'cover',
+        backgroundColor: 'rgba(0, 0, 0, 0.1)'
+    }
+
+    console.log(background)
     return (
-        <section className='single-movie'>
+        <section className='single-movie' style={ background }>
             <div className='movie-stats'>
                 <img className='poster' src={props.singleMovie.poster_path} alt={`${props.singleMovie.title} Movie Poster`} />
                 <h1>{props.singleMovie.title}</h1>
                 <h4>{props.singleMovie.release_date}&nbsp;&nbsp; | &nbsp;&nbsp;{props.singleMovie.average_rating}</h4> 
-                 {/* | {props.singleMovie.runtime} */}
+                <h4>Runtime: {`${props.singleMovie.runtime} minutes`}</h4>
             </div>
             <div className='movie-details'>
-                {/* <p>Genre: {props.genres}</p> */}
-                {/* <p>Description: {props.overview}</p> */}
-                {/* <p>Budget: ${props.budget}</p> */}
-                <p>Release Date: {props.singleMovie.release_date}</p>
+                <p>Genre: {props.singleMovie.genres}</p>
+                <p>Description: {props.singleMovie.overview}</p>
+
                 <div className='backArrow'>
                     <img role='button' src={backArrow} alt='back button' width='75vw' height='75vh' onClick={ props.goBack } />
                 </div>
